# frontend/openapi.json
{"openapi":"3.1.0","info":{"title":"Interview Coach API","version":"0.1.0"},"paths":{"/{url}":{"options":{"summary":"Handle Options","operationId":"handle_options__url__options","parameters":[{"name":"url","in":"path","required":true,"schema":{"title":"Url"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/ping":{"get":{"summary":"Ping Pong","operationId":"ping_pong_ping_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/user/register":{"post":{"tags":["user"],"summary":"Register","operationId":"register_user_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshAccessToken"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/user/refresh":{"post":{"tags":["user"],"summary":"Refresh Token Route","operationId":"refresh_token_route_user_refresh_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_refresh_token_route_user_refresh_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshAccessToken"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/user/token":{"post":{"tags":["user"],"summary":"Login For Access Token","operationId":"login_for_access_token_user_token_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_login_for_access_token_user_token_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshAccessToken"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/user/self":{"get":{"tags":["user"],"summary":"Get Self","operationId":"get_self_user_self_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/user":{"put":{"tags":["user"],"summary":"Update User","operationId":"update_user_user_put","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserUpdate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]},"delete":{"tags":["user"],"summary":"Delete User","operationId":"delete_user_user_delete","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Delete User User Delete"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/user/all":{"get":{"tags":["user"],"summary":"Get All Users","operationId":"get_all_users_user_all_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/UserResponse"},"type":"array","title":"Response Get All Users User All Get"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/user/email":{"get":{"tags":["user"],"summary":"Get User By Email","operationId":"get_user_by_email_user_email_get","parameters":[{"name":"email","in":"query","required":true,"schema":{"type":"string","title":"Email"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/interviews":{"post":{"tags":["interviews"],"summary":"Create Interview","operationId":"create_interview_interviews_post","security":[{"OAuth2PasswordBearer":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InterviewCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InterviewCreateResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["interviews"],"summary":"List Interviews","operationId":"list_interviews_interviews_get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"offset","in":"query","required":false,"schema":{"type":"integer","minimum":0,"default":0,"title":"Offset"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"default":20,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InterviewListResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/interviews/{session_id}/messages":{"post":{"tags":["interviews"],"summary":"Post Interview Message","operationId":"post_interview_message_interviews__session_id__messages_post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"session_id","in":"path","required":true,"schema":{"type":"integer","title":"Session Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InterviewMessageCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InterviewReplyResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/interviews/{session_id}/complete":{"post":{"tags":["interviews"],"summary":"Complete Interview","operationId":"complete_interview_interviews__session_id__complete_post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"session_id","in":"path","required":true,"schema":{"type":"integer","title":"Session Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InterviewReportOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/interviews/progress/summary":{"get":{"tags":["interviews"],"summary":"Interview Progress Summary","operationId":"interview_progress_summary_interviews_progress_summary_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProgressSummaryResponse"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/interviews/{session_id}":{"get":{"tags":["interviews"],"summary":"Get Interview","operationId":"get_interview_interviews__session_id__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"session_id","in":"path","required":true,"schema":{"type":"integer","title":"Session Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InterviewDetailResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Body_login_for_access_token_user_token_post":{"properties":{"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"}},"type":"object","required":["username","password"],"title":"Body_login_for_access_token_user_token_post"},"Body_refresh_token_route_user_refresh_post":{"properties":{"refresh_token":{"type":"string","title":"Refresh Token"}},"type":"object","required":["refresh_token"],"title":"Body_refresh_token_route_user_refresh_post"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"InterviewCreate":{"properties":{"track":{"$ref":"#/components/schemas/InterviewTrackEnum"},"level":{"$ref":"#/components/schemas/InterviewLevelEnum"},"mode":{"$ref":"#/components/schemas/InterviewModeEnum","default":"technical"},"stress":{"type":"boolean","title":"Stress","default":false},"max_turns":{"type":"integer","maximum":30.0,"minimum":1.0,"title":"Max Turns","default":8}},"type":"object","required":["track","level"],"title":"InterviewCreate"},"InterviewCreateResponse":{"properties":{"session":{"$ref":"#/components/schemas/InterviewSessionOut"},"assistant_message":{"$ref":"#/components/schemas/InterviewMessageOut"}},"type":"object","required":["session","assistant_message"],"title":"InterviewCreateResponse"},"InterviewDetailResponse":{"properties":{"session":{"$ref":"#/components/schemas/InterviewSessionOut"},"messages":{"items":{"$ref":"#/components/schemas/InterviewMessageOut"},"type":"array","title":"Messages"}},"type":"object","required":["session","messages"],"title":"InterviewDetailResponse"},"InterviewLevelEnum":{"type":"string","enum":["Junior","Middle","Senior"],"title":"InterviewLevelEnum"},"InterviewListItem":{"properties":{"session":{"$ref":"#/components/schemas/InterviewSessionOut"},"overall_score":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Overall Score"}},"type":"object","required":["session"],"title":"InterviewListItem"},"InterviewListResponse":{"properties":{"items":{"items":{"$ref":"#/components/schemas/InterviewListItem"},"type":"array","title":"Items"},"total":{"type":"integer","title":"Total"}},"type":"object","required":["items","total"],"title":"InterviewListResponse"},"InterviewMessageCreate":{"properties":{"content":{"type":"string","maxLength":50000,"minLength":1,"title":"Content"}},"type":"object","required":["content"],"title":"InterviewMessageCreate"},"InterviewMessageOut":{"properties":{"id":{"type":"integer","title":"Id"},"role":{"type":"string","title":"Role"},"content":{"type":"string","title":"Content"},"created_at":{"type":"string","title":"Created At"}},"type":"object","required":["id","role","content","created_at"],"title":"InterviewMessageOut"},"InterviewMessage_OVN":{"properties":{"id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Id"},"role":{"type":"string","title":"Role"},"content":{"type":"string","title":"Content"},"raw_llm_payload":{"anyOf":[{},{"type":"null"}],"title":"Raw Llm Payload"},"created_at":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Created At"}},"type":"object","required":["role","content","created_at"],"title":"InterviewMessage_OVN"},"InterviewModeEnum":{"type":"string","enum":["technical","hr","mixed"],"title":"InterviewModeEnum"},"InterviewReplyResponse":{"properties":{"session":{"$ref":"#/components/schemas/InterviewSessionOut"},"assistant_message":{"$ref":"#/components/schemas/InterviewMessageOut"},"interview_finished":{"type":"boolean","title":"Interview Finished"}},"type":"object","required":["session","assistant_message","interview_finished"],"title":"InterviewReplyResponse"},"InterviewReportOut":{"properties":{"id":{"type":"integer","title":"Id"},"session_id":{"type":"integer","title":"Session Id"},"summary_text":{"type":"string","title":"Summary Text"},"scores":{"additionalProperties":true,"type":"object","title":"Scores"},"recommendations":{"items":{},"type":"array","title":"Recommendations"},"weak_areas":{"items":{},"type":"array","title":"Weak Areas"},"study_plan":{"items":{},"type":"array","title":"Study Plan"},"created_at":{"type":"string","title":"Created At"}},"type":"object","required":["id","session_id","summary_text","scores","recommendations","weak_areas","study_plan","created_at"],"title":"InterviewReportOut"},"InterviewReport_VEI":{"properties":{"id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Id"},"summary_text":{"type":"string","title":"Summary Text"},"scores":{"title":"Scores"},"recommendations":{"title":"Recommendations"},"weak_areas":{"title":"Weak Areas"},"study_plan":{"title":"Study Plan"},"created_at":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Created At"}},"type":"object","required":["summary_text","scores","recommendations","weak_areas","study_plan","created_at"],"title":"InterviewReport_VEI"},"InterviewSessionOut":{"properties":{"id":{"type":"integer","title":"Id"},"track":{"type":"string","title":"Track"},"level":{"type":"string","title":"Level"},"mode":{"type":"string","title":"Mode"},"stress":{"type":"boolean","title":"Stress"},"status":{"type":"string","title":"Status"},"max_turns":{"type":"integer","title":"Max Turns"},"created_at":{"type":"string","title":"Created At"},"completed_at":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Completed At"}},"type":"object","required":["id","track","level","mode","stress","status","max_turns","created_at"],"title":"InterviewSessionOut"},"InterviewSession_JFQ":{"properties":{"id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Id"},"track":{"type":"string","title":"Track"},"level":{"type":"string","title":"Level"},"mode":{"type":"string","title":"Mode"},"stress":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Stress","default":false},"status":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Status","default":"active"},"max_turns":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Max Turns","default":8},"extra_metadata":{"anyOf":[{},{"type":"null"}],"title":"Extra Metadata"},"created_at":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Created At"},"completed_at":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Completed At"},"interviewmessages":{"anyOf":[{"items":{"$ref":"#/components/schemas/InterviewMessage_OVN"},"type":"array"},{"type":"null"}],"title":"Interviewmessages"},"interviewreports":{"anyOf":[{"items":{"$ref":"#/components/schemas/InterviewReport_VEI"},"type":"array"},{"type":"null"}],"title":"Interviewreports"}},"type":"object","required":["track","level","mode","created_at"],"title":"InterviewSession_JFQ"},"InterviewTrackEnum":{"type":"string","enum":["Backend","Frontend","QA","DevOps"],"title":"InterviewTrackEnum"},"ProgressSummaryResponse":{"properties":{"sessions_completed":{"type":"integer","title":"Sessions Completed"},"average_overall_score":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Average Overall Score"},"common_weak_areas":{"items":{"type":"string"},"type":"array","title":"Common Weak Areas"}},"type":"object","required":["sessions_completed","average_overall_score","common_weak_areas"],"title":"ProgressSummaryResponse"},"RefreshAccessToken":{"properties":{"refresh_token":{"type":"string","title":"Refresh Token"},"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type","default":"bearer"}},"type":"object","required":["refresh_token","access_token"],"title":"RefreshAccessToken"},"UserCreate":{"properties":{"name":{"type":"string","title":"Name"},"patronymic":{"type":"string","title":"Patronymic"},"surname":{"type":"string","title":"Surname"},"date_of_birth":{"type":"string","title":"Date Of Birth"},"email":{"type":"string","title":"Email"},"password":{"type":"string","title":"Password"}},"type":"object","required":["name","patronymic","surname","date_of_birth","email","password"],"title":"UserCreate"},"UserResponse":{"properties":{"id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Id"},"name":{"type":"string","title":"Name"},"patronymic":{"type":"string","title":"Patronymic"},"surname":{"type":"string","title":"Surname"},"email":{"type":"string","title":"Email"},"date_of_birth":{"type":"string","title":"Date Of Birth"},"role":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Role","default":"USER"},"interviewsessions":{"anyOf":[{"items":{"$ref":"#/components/schemas/InterviewSession_JFQ"},"type":"array"},{"type":"null"}],"title":"Interviewsessions"}},"type":"object","required":["name","patronymic","surname","email","date_of_birth"],"title":"UserResponse"},"UserUpdate":{"properties":{"email":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Email"},"name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Name"},"patronymic":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Patronymic"},"surname":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Surname"},"date_of_birth":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Date Of Birth"}},"type":"object","title":"UserUpdate"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"/user/token"}}}}}}